# src/main/resources/openapi.yaml
openapi: 3.1.0
info:
  title: to do list - OpenAPI 3.1
  description: |-
    This is a to do list server based on roadmap.sh's project specifications.
    Find out more at: 
      https://roadmap.sh/projects/todo-list-api
    
    Useful links:
    - [to do list repository](https://github.com/jocelynzaman/Roadmap-Todo-List-API)

  contact:
    email: jocelynzaman@gmail.com
  license:
    name: MIT
  version: 1.0.0
servers:
  - url: localhost:8094/
tags:
  - name: user authentication
    description: Sign up or log in to access to do list
paths:
  /register:
    post:
      tags:
        - user authentication
      summary: Register a new user
      description: Register a new user to access to do list
      operationId: registerUser
      requestBody:      
        description: Create a new user
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRegistration'
        required: true
      responses:
        '201':
          description: Successful 
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Token'
        '400':
          description: Missing or invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegistrationException'
components:
  schemas:
    UserRegistration:
      type: object
      properties:
        name:
          type: string
          examples: ["Dipper Pines"]
        email:
          type: string
          examples: ["Dipper.Pines@gmail.com"]
        password:
          type: string
          examples: ["gravity"]
    Token:
      type: object
      properties:
        token:
          type: string
          examples: ["EewQoOTnJ7M6EuElcJZFgtssu"]
    RegistrationException:
      type: object
      properties:
        message:
          type: string
  requestBodies:
    UserRegistration:
      description: User object that needs to be added to the application
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/UserRegistration'